generator/requirements.txt
tree-sitter==0.24.0
tree-sitter-cpp==0.23.4
markdown-it-py==3.0.0
pygments==2.19.1
